requirements.txt
Flask==1.1.2
gunicorn==19.9.0
itsdangerous==2.0.1
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==2.0.2
numpy>=1.20.3
tensorflow-cpu


